feat: nuxt hooks
