+++ examples/nuxt-simple/nuxt.config.js
@@ -3,7 +3,7 @@ module.exports = {
    ** Headers of the page
    */
   head: {
-    title: "my-nuxt-project",
+    title: "Vue Query Nuxt Example",
     meta: [
       { charset: "utf-8" },
       { name: "viewport", content: "width=device-width, initial-scale=1" },
