+++ examples/nuxt-simple/package.json
@@ -13,7 +13,7 @@
     "precommit": "npm run lint"
   },
   "dependencies": {
-    "@nuxtjs/composition-api": "^0.24.2",
+    "@nuxtjs/composition-api": "^0.24.4",
     "nuxt": "^2.0.0",
     "vue-query": "^1.5.1"
   },
