+++ package.json
@@ -43,14 +43,19 @@
   },
   "peerDependencies": {
     "@vue/composition-api": "^1.0.0-rc.1",
+    "@nuxtjs/composition-api": "^0.24.4",
     "vue": "^2.0.0 || >=3.0.0"
   },
   "peerDependenciesMeta": {
     "@vue/composition-api": {
       "optional": true
+    },
+    "@nuxtjs/composition-api": {
+      "optional": true
     }
   },
   "devDependencies": {
+    "@nuxtjs/composition-api": "^0.24.4",
     "@rollup/plugin-node-resolve": "^11.2.0",
     "@types/jest": "^26.0.21",
     "@typescript-eslint/eslint-plugin": "^4.16.1",
@@ -74,7 +79,7 @@
     "ts-node": "^9.1.1",
     "typescript": "^4.2.3",
     "vue": "^3.0.7",
-    "vue2": "npm:vue@2",
-    "vue-jest": "^5.0.0-alpha.8"
+    "vue-jest": "^5.0.0-alpha.8",
+    "vue2": "npm:vue@2"
   }
 }
