+++ src/ssr/index.ts
@@ -1,1 +1,4 @@
 export { hydrate, dehydrate } from "./hydration";
+
+export { useNuxtQueryProvider } from "./useNuxtQueryProvider";
+export { useNuxtDehydrate } from "./useNuxtDehydrate";
